courses/ml-for-engineers/production/module-7/projects/cricket-shot-recognizer/requirements.txt
tensorflow==2.13.0
streamlit==1.28.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-learn==1.3.0
